Add mealybug-tearoom-tests to test-suite
This includes test for DMG, CGB revision C and CGB revision D, but only
DMG tests are beign run for now. That is because it is not clear which
revisions the various Game Boy Color emulators are emulating, and the
ones that I checked (SameBoy and Beaten Dying Moon) were emulating the
revision E, whose expected results are not included in the test suite.