Emmy: relie on the enviroment to provide ChromeDriver
Emmy is a browser based emulator, and uses a WebDriver to capture the
emulator screen. Before this change, emmy was downloading and installing
the ChromeDriver, but that need to match with the version of Chrome
installed on the current enviroment, which wasn't.

To fix that, now Emmy expects the enviroment to already have the correct
ChromeDriver version installed. On Github Actions, we relie on a
third-party action to install the correct version of ChromeDriver.