Handle KeyboardInterrupt a little better
If you pressed Ctrl-C while a test is running, that test would fail, but
the test suite would continue running.

With this, the test suite will exit on Ctrl-C. This is still not very
gracefull (it still keeps the current emulator open after exiting, for
example), but makes it more easy to kill the program.